+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Implement `Attr#cloneNode()`
 * Implement `Element#getAttributeNode()`, `Element#getAttributeNodeNS()`, and
   `Element#hasAttributes()`
+* Implement `Text#wholeText`
 
 # domino 2.0.3 (12 Jul 2018)
 * Define `blur()`, `focus()` and `forceSpellCheck()` on `HTMLElement` (#125)